Fixing warning from previous cleanup on ModelGraph.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@545 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/Models/PolyhedronModel.cpp
@@ -162,7 +162,7 @@ PolyhedronModel::BuildWired(const PtVector& _points, const TriVector& _tris, con
     int trir = (*eit).property()[1];
 
     if(tril == -1 || trir == -1 || 1-fabs(_norms[tril] * _norms[trir]) > 1e-3) {
-      GLint id[2] = {(*eit).source(), (*eit).target()};
+      GLint id[2] = {(int)(*eit).source(), (int)(*eit).target()};
       glDrawElements(GL_LINES, 2, GL_UNSIGNED_INT, id);
     }
   }
